% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfseries-cigits.R
\docType{data}
\name{vf.cigts}
\alias{vf.cigts}
\title{Combined Visual Field Series for General Progression Method}
\format{A data frame sample for CIGTS progression method, which includes visual field related measurement for two eyes, each with 10 follow-ups. Rows represent the single measurements.}
\source{
\describe{
  \item{eyeid}{eyeid, labeled as 1,2... for different eyes.}
  \item{yearsfollowed}{follow-up years. The minimum measurements /rows for one eye is 5.}
  \item{tdp1-tdp54}{52 total deviation probability, or 'tdp' measurements. The minimum measurements, or rows for one eye is 5.}
  ...
}
}
\usage{
data(vf.cigts)
}
\description{
Data
}
\examples{
data(vf.cigts)
colnames(vf.cigts)
progression.cigts(vf.cigts)
progression.cigts(vf.cigts[vf.cigts$eyeid == 1,])
progression.cigts(vf.cigts[vf.cigts$eyeid == 2,])
}
\keyword{datasets}
