% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfseries-plr-nouri-2012.R
\docType{data}
\name{vf.plr.nouri.2012}
\alias{vf.plr.nouri.2012}
\title{Combined Visual Field Series for General Progression Method}
\format{A data frame sample for Pointwise Linear Regression (PLR) method according to Nouri-Mahdavi 2012 progression, which includes visual field related measurement for two eyes, each with 10 follow-ups. Rows represent the single measurements.}
\source{
\describe{
  \item{eyeid}{eyeid, labeled as 1,2... for different eyes}
  \item{yearsfollowed}{follow-up years. The minimum measurements, or rows, for one eye is 3}
  \item{td1-td54}{52 total deviation, or 'td' measurements. The minimum measurements, or rows, for one eye is 3}
  ...
}
}
\usage{
data(vf.plr.nouri.2012)
}
\description{
Data
}
\examples{
data(vf.plr.nouri.2012)
colnames(vf.plr.nouri.2012)
progression.plr.nouri.2012(vf.plr.nouri.2012)
progression.plr.nouri.2012(vf.plr.nouri.2012[vf.plr.nouri.2012$eyeid == 1,])
progression.plr.nouri.2012(vf.plr.nouri.2012[vf.plr.nouri.2012$eyeid == 2,])
}
\keyword{datasets}
