% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfseries-schell2014.R
\docType{data}
\name{vf.schell2014}
\alias{vf.schell2014}
\title{Combined Visual Field Series for General Progression Method}
\format{A data frame sample for progression method by Schell et al. 2014, which includes visual field related measurement for two eyes, each with 10 follow-ups. Rows represent the single measurements.}
\source{
\describe{
  \item{eyeid}{eyeid, labeled as 1,2... for different eyes.}
  \item{md}{mean deviation measurements. The minimum measurements, or rows, for one eye is 4.}
  ...
}
}
\usage{
data(vf.schell2014)
}
\description{
Data
}
\examples{
data(vf.schell2014)
colnames(vf.schell2014)
progression.schell2014(vf.schell2014)
progression.schell2014(vf.schell2014[vf.schell2014$eyeid == 1,])
progression.schell2014(vf.schell2014[vf.schell2014$eyeid == 2,])
}
\keyword{datasets}
