% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vfseries-vfi.R
\docType{data}
\name{vf.vfi}
\alias{vf.vfi}
\title{Combined Visual Field Series for General Progression Method}
\format{A data frame for CIGTS progression example, which includes visual field related measurement for two eyes each with 10 follow-ups.}
\source{
\describe{
  \item{eyeid}{eyeid, labeled as 1,2... for different eye groups.}
  \item{yearsfollowed}{follow-up years. The minimum measurements, or rows, for one eye is 3.}
  \item{vfi}{visual field index. The minimum measurements, or rows, for one eye is 3.}
  ...
}
}
\usage{
data(vf.vfi)
}
\description{
Data
}
\examples{
data(vf.vfi)
colnames(vf.vfi)
progression.vfi(vf.vfi)
progression.vfi(vf.vfi[vf.vfi$eyeid == 1,])
progression.vfi(vf.vfi[vf.vfi$eyeid == 2,])
}
\keyword{datasets}
